enhancement: refactor contests calls (#3717)
* enhancement: refactor contests calls

* enhancement: fix build error

* enhancement: import hash from contracts

* enhancement: clean some code
