+++ packages/react-app-revamp/lib/contests/constants/index.ts
@@ -1,1 +1,4 @@
 export const CONTESTS_FEATURE_COUNT = 8;
+
+export const FEATURED_CONTEST_COLUMNS =
+  "created_at, start_at, end_at, address, author_address, network_name, vote_start_at, featured, type, submissionMerkleRoot, votingMerkleRoot, voting_requirements, submission_requirements";
